### 2015-09-21 4c46830
Remove unused option from as_strings
