+++ R/utils.R
@@ -1,5 +1,5 @@
 # Take results of readLines and turn it into a character vector of length 1
-as_string <- function(x, collapse = " ") {
+as_string <- function(x) {
   x %>%
     str_c(collapse = "\n") %>%
     NLP::as.String()
