### 2015-09-22 befed49
Much faster ratio of matches
