### 2015-09-22 ef66420
Function to rehash documents
Includes changes to `tokenize` to keep the functions similar, and improvements to vignette.
