+++ R/tokenize.R
@@ -4,17 +4,19 @@
 #' \code{\link{TextReuseCorpus}}, this function recomputes the tokens and hashes
 #' with the functions specified.
 #'
-#' @param x A \code{\link{TextReuseTextDocument}} or \code{\link{TextReuseCorpus}}.
+#' @param x A \code{\link{TextReuseTextDocument}} or
+#'   \code{\link{TextReuseCorpus}}.
 #' @param tokenizer A function to split the text into tokens. See
 #'   \code{\link{tokenizers}}.
 #' @param ... Arguments passed on to the \code{tokenizer}.
 #' @param hash_func A function to hash the tokens. See
-#'   \code{\link{hash_string}}.
+#'   \code{\link{hash_string}}, \code{\link{minhash_generator}}.
 #' @param keep_tokens Should the tokens be saved in the document that is
 #'   returned or discarded?
 #' @param keep_text Should the text be saved in the document that is returned or
 #'   discarded?
-#' @return The modified \code{\link{TextReuseTextDocument}} or \code{\link{TextReuseCorpus}}.
+#' @return The modified \code{\link{TextReuseTextDocument}} or
+#'   \code{\link{TextReuseCorpus}}.
 #' @export
 tokenize <- function(x, tokenizer, ..., hash_func = hash_string,
                      keep_tokens = FALSE, keep_text = TRUE) {
@@ -26,7 +28,9 @@ tokenize.TextReuseTextDocument <- function(x, tokenizer, ...,
                                            hash_func = hash_string,
                                            keep_tokens = TRUE,
                                            keep_text = TRUE) {
-  assert_that(has_content(x))
+  assert_that(has_content(x),
+              is.function(tokenizer),
+              is.function(hash_func))
   x$tokens <- tokenizer(x$content, ...)
   x$hashes <- hash_func(x$tokens)
   if (!keep_tokens) x$tokens <- NULL
