+++ NAMESPACE
@@ -27,6 +27,8 @@ S3method(print,TextReuseCorpus)
 S3method(print,TextReuseTextDocument)
 S3method(ratio_of_matches,TextReuseTextDocument)
 S3method(ratio_of_matches,default)
+S3method(rehash,TextReuseCorpus)
+S3method(rehash,TextReuseTextDocument)
 S3method(tokenize,TextReuseCorpus)
 S3method(tokenize,TextReuseTextDocument)
 S3method(tokens,TextReuseCorpus)
@@ -43,6 +45,7 @@ export(TextReuseTextDocument)
 export(content)
 export(filenames)
 export(has_content)
+export(has_tokens)
 export(hash_string)
 export(hashes)
 export(is.TextReuseCorpus)
@@ -61,6 +64,7 @@ export(minhash_generator)
 export(pairwise_candidates)
 export(pairwise_compare)
 export(ratio_of_matches)
+export(rehash)
 export(tokenize)
 export(tokenize_ngrams)
 export(tokenize_sentences)

[…]
