### 2015-09-22 15c2e1b
Fix subsetting bug
