### 2015-09-22 #20 @lmullen
Just a note to whoever ends up reviewing this package: I've made some performance improvements/bug fixes, but these are all on master now (with tags).
