### 2015-09-22 2450807
Add progress bar to LSH
