### 2015-09-22 ac7049c
Add has_hashes, and take exported functions outside of utils
