+++ man/TextReuseTextDocument.Rd
@@ -1,8 +1,9 @@
 % Generated by roxygen2 (4.1.1): do not edit by hand
-% Please edit documentation in R/TextReuseTextDocument.R, R/utils.R
+% Please edit documentation in R/TextReuseTextDocument.R
 \name{TextReuseTextDocument}
 \alias{TextReuseTextDocument}
 \alias{has_content}
+\alias{has_hashes}
 \alias{has_tokens}
 \alias{is.TextReuseTextDocument}
 \title{TextReuseTextDocument}
@@ -16,6 +17,8 @@ is.TextReuseTextDocument(x)
 has_content(x)
 
 has_tokens(x)
+
+has_hashes(x)
 }
 \arguments{
 \item{text}{A character vector containing the text of the document. This
