+++ NAMESPACE
@@ -45,6 +45,7 @@ export(TextReuseTextDocument)
 export(content)
 export(filenames)
 export(has_content)
+export(has_hashes)
 export(has_tokens)
 export(hash_string)
 export(hashes)
