### 2015-09-23 2461c0e
Allow loading documents without tokenizing/hashing
