### 2015-09-23 2d4d8c8
Permit NULL tokenizer in TextReuseCorpus
