### 2015-09-26 a8d937e
Keep lsh hashes distinct by including band
