### 2015-09-26 f9d3555
Fix loading message when not tokenizing.
