### 2015-10-16 #20 @noamross
# General Comments

👏👏👏

This package is so f*ing elegant.  It's classy, really goddam classy and I barely have anything to add and kind of want to delete everything on my hard drive in response.

Some things I really, really like:

-  Right-sized: The package implements a few awesome methods for a particular task,
   and doesn't try too much else.
-  Excellent test coverage
-  Constant use of assertions
-  Well organized, readable code
-  Just the right amount of S3
-  Well-written, easy to follow, multiple short vignettes
-  `progress = interactive()` as default for all long-running functions.

I'm wasn't familiar with the particular methods described here but they were easy to learn from the vignettes and I was able to apply them to a corpus of 3000 documents in no time.

The package is also a testament to how much recent tooling such as `testthat`, `assertthat`, `Rcpp`, and such have improved the quality of R packages.

I really couldn't find anything in the code to complain about.  The following is mostly a bunch of tiny feature and documentation requests.

Great, job, @lmullen!

# Tests

- Tests run successfully on OSX Yosemite, R 3.2.2
- All code in README, vignettes and examples ran as expected

# Documentation

-  In the intro vingettes, there isn't text specifically explaining the `meta()` function

-  `?textreuse` only defines the datasets.  I find having a short description of the package and pointers to vignettes or key functions helpful here.

-  The documentation discusses how the user may use or write other tokenizing, hash, or dissimilarity functions.  It would be helpful for examples to include a case or two of using a function from another package.  Also, if you hope for others to do so, you might provide a template in a `CONTRIBUTING.md` file.

-  In the README as well was the minhashing vignette you mention that pairwise comparisons grow exponentially with the size of the corpus.  I think the term *geometrically* is correct here. (Can you see that I'm really reaching for things?.

-  In `minhash.Rmd`, line 63: "same four", but this identifies only 3.
-  In `rehash.R`, line 5: tokenens --> tokens
-  In `rehash.R`, line 5: which --> when
-  In `lsh_compare.R`, line 6:  "each of the documents" --> "each of the candidate pairs"

-  I recommend adding a [code of conduct](https://github.com/ropensci/packaging_guide#-code-of-conduct) to the package. (`devtools:::use_code_of_conduct()`)

# Functionality

-   I would like to be able to create a TextReuseCorpus from a character vector rather than just a directory of files, as I commonly am loading a bunch of other files (e.g., html) and processing them to extract just the relevant text.  In this case one would have to provide a vector of ids or use the vector names as ids, which could default to just using the index number as id.

-  When loading in a large corpus of documents, I got the error `Error: n not less than length(words)`, which was ultimately traced to `assert_that(n < length(words))` in `tokenize_ngrams()`.  I found this was because I had a very short document, but also because the document was all whitespace characters. I wonder if it would make sense to check for empty documents.

-  Perhaps it makes sense to give pairwise candidate dataframes their own class for assertion testing, rather than testing on column names, as in `lsh_compare()`.

- One needs to use the same minhashing function when adding new documents to a hash.  Perhaps the hash could store the minhash function and bands used to create it, and check against this for new documents.
