### 2015-10-16 #20 @lmullen
@noamross Thanks so much for this very thorough review. I really appreciate the level of detail of your review, which gives me a lot more confidence in the package. I'll respond point by point as necessary as I work on implementing your suggestions. But in general, I think everything that you have suggested is worth doing, and I plan on implementing them before sending the package to CRAN. In particular, I like the idea of giving the pairwise data frames a class: much more elegant. Also, I ran into the `n not less than length(words)` error when using this for my own analysis. I'll have think about the best way to either document or work around that: what I've been doing is filtering the corpus with `wordcount()`.

One other thing. I've noticed that a performance problem using the hash package (ugh---so many kinds of hashes in this package!). With a corpus of 71K documents, the `lsh_compare()` function ran for 3 days, got to about 30% completion, and wouldn't budge from there before I quit. The problem is that while hash tables are very fast for reading data, they don't seem to have any advantage in writing data. In particular, I don't think the implementation of hash tables in the hash package lets you preallocate the size of the hash table, which I couldn't do anyway because there is no way of knowing in advance how many collisions there will be with the hashes. Maybe I'm doing something naive with the hash table and just don't know it. But I've got an implementation (external to the package, [example here](https://github.com/lmullen/civil-procedure-codes/blob/master/016-manual-lsh.md)) that uses dplyr to calculate the matches. It took less than 90 minutes for the 71K document corpus. This will entail dropping the dependency on the hash package and adding a dependency on dplyr. I think that will be simpler for users, since the hash table is an unfamiliar data structure. And while dplyr is a somewhat heavier dependency, in the long run it will permit using a database backend so that the package can be used with out of memory data.

Thanks, Noam!
