### 2015-10-16 #20 @noamross
I'm happy to review a new implementation.
