### 2015-10-16 144ca42
Spell fixes in docs
(And regenerate the docs)
