### 2015-10-16 lmullen/textreuse#44 @lmullen
Create corpus from character vector (or possibly a list)
From [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/20):

> I would like to be able to create a TextReuseCorpus from a character vector rather than just a directory of files, as I commonly am loading a bunch of other files (e.g., html) and processing them to extract just the relevant text. In this case one would have to provide a vector of ids or use the vector names as ids, which could default to just using the index number as id.
