### 2015-10-16 lmullen/textreuse#45 @lmullen
Better handling of documents where number of words is less than n in n-grams
From [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/20):

> When loading in a large corpus of documents, I got the error Error: n not less than length(words), which was ultimately traced to assert_that(n < length(words)) in tokenize_ngrams(). I found this was because I had a very short document, but also because the document was all whitespace characters. I wonder if it would make sense to check for empty documents.

Several things to do here:

- [x] Error message should be more descriptive.
- [x] Show in the vignette how to filter short documents.
