### 2015-10-16 lmullen/textreuse#46 @lmullen
Replace checks for columns with class identifier
From [rOpenSci onboarding](https://github.com/ropensci/onboarding/issues/20)

> Perhaps it makes sense to give pairwise candidate dataframes their own class for assertion testing, rather than testing on column names, as in lsh_compare().
