### 2015-10-17 15aaa01
Skip on appveyor function in testthat now
