+++ tests/testthat/test-wordcount.R
@@ -12,11 +12,9 @@ test_that("counts words correctly for different classes", {
 
 test_that("counts words for a corpus", {
   skip_on_appveyor()
-  if (.Platform$OS.type == "unix") {
-    wc <- wordcount(corpus)
-    expect_true(!is.null(names(wc)))
-    wc <- unname(wc)
-    expect_equal(wc, c(729, 217, 790))
-  }
+  wc <- wordcount(corpus)
+  expect_true(!is.null(names(wc)))
+  wc <- unname(wc)
+  expect_equal(wc, c(729, 217, 790))
 })
 
