+++ DESCRIPTION
@@ -21,7 +21,7 @@ Imports: assertthat (>= 0.1),
     NLP (>= 0.1.8),
     Rcpp (>= 0.12.0),
     stringr (>= 1.0.0)
-Suggests: testthat (>= 0.10.0),
+Suggests: testthat (>= 0.11.0),
     knitr (>= 1.11),
     rmarkdown (>= 0.8)
 LinkingTo: Rcpp,
