### 2015-10-17 2cd098f
Explanation of meta function
