### 2015-10-17 d60c8a5
Demonstrate tokenizing functions (with a poem)
