### 2015-10-17 dd2abe3
Geometrically, not exponentially
