+++ vignettes/minhash.Rmd
@@ -9,7 +9,7 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-Performing pairwise comparisons in a corpus is time-consuming because the number of comparisons grows exponentially with the size of the corpus. Most of those comparisons, furthermore, are unnecessary because they do not result in matches. The combination of minhash and locality-sensitive hashing (LSH) seeks to solve these problems. They make it possible to compute possible matches only once for each document, so that the cost of computation grows linearly rather than exponentially. This vignette explains how to use the minhash and locality-sensitive hashing functions in this package. For an explanation of why they work, see Jure Leskovec, Anand Rajaraman, and Jeff Ullman, *[Mining of Massive Datasets](http://www.mmds.org/#book)* (Cambridge University Press, 2011), ch. 3. (This [blog post](http://matthewcasperson.blogspot.com/2013/11/minhash-for-dummies.html) is a more succinct explanation.)
+Performing pairwise comparisons in a corpus is time-consuming because the number of comparisons grows geometrically with the size of the corpus. Most of those comparisons, furthermore, are unnecessary because they do not result in matches. The combination of minhash and locality-sensitive hashing (LSH) seeks to solve these problems. They make it possible to compute possible matches only once for each document, so that the cost of computation grows linearly rather than exponentially. This vignette explains how to use the minhash and locality-sensitive hashing functions in this package. For an explanation of why they work, see Jure Leskovec, Anand Rajaraman, and Jeff Ullman, *[Mining of Massive Datasets](http://www.mmds.org/#book)* (Cambridge University Press, 2011), ch. 3. (This [blog post](http://matthewcasperson.blogspot.com/2013/11/minhash-for-dummies.html) is a more succinct explanation.)
 
 We begin by creating a minhash function. A minhash function converts tokenized text into a set of hash integers, then selects the minimum value. This is the equivalent of randomly selecting a token. The function then does the same thing repeatedly with different hashing functions, in effect selecting `n` random shingles. The additional hashing functions come from a bitwise XOR with random integers. That is why the `minhash_generator()` accepts a seed, so that we can re-create the same minhash function again. In other words, a minhash function converts a set of tokens of any length into `n` randomly selected and hashed tokens.
 
