+++ man/rehash.Rd
@@ -22,7 +22,7 @@ Given a \code{\link{TextReuseTextDocument}} or a
 \code{\link{TextReuseCorpus}}, this function recomputes the hashes with the
 function specified. This implies that you have retained the tokens with the
 \code{keep_tokens = TRUE} parameter. This function is intended to be used
-which changing from keeping only minhashes of the tokens to keeping hashes of
+when changing from keeping only minhashes of the tokens to keeping hashes of
 all the tokens. See the vignette on minhash/LSH.
 }
 
