+++ vignettes/minhash.Rmd
@@ -60,7 +60,7 @@ candidates <- lsh_candidates(buckets)
 candidates
 ```
 
-Notice that LSH has identified the same four pairs of documents as potential matches that we found with pairwise comparisons, but did so much faster. But we do not have similarity scores; we only know that these documents are likely to have Jaccard similarity scores above the `r round(lsh_threshold(h = 240, b = 80), 3)` threshold.
+Notice that LSH has identified the same three pairs of documents as potential matches that we found with pairwise comparisons, but did so much faster. But we do not have similarity scores; we only know that these documents are likely to have Jaccard similarity scores above the `r round(lsh_threshold(h = 240, b = 80), 3)` threshold.
 
 We could compute Jaccard similarities using the minhashes, or we could retokenize and hash only the candidate documents. We will do the later, using the `lsh_subset()` function to get just the documents that we want and the `rehash()` function to compute hashes for all of the tokens in the subset corpus. (If we had not keep the tokens with `keep_tokens = TRUE` above, we would have to retokenize with `tokenize()` instead.)
 
