+++ README.Rmd
@@ -87,3 +87,6 @@ If you use this package for scholarly research, I would appreciate a citation.
 citation("textreuse")
 ```
 
+### Contributing
+
+Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
