+++ README.md
@@ -108,3 +108,7 @@ citation("textreuse")
 #>     url = {https://github.com/lmullen/textreuse},
 #>   }
 ```
+
+### Contributing
+
+Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
