+++ README.Rmd
@@ -87,6 +87,8 @@ If you use this package for scholarly research, I would appreciate a citation.
 citation("textreuse")
 ```
 
-### Contributing
+### Contributing and acknowledgements
 
 Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms.
+
+Thanks to [Noam Ross](http://www.noamross.net/) for his thorough [peer review](https://github.com/ropensci/onboarding/issues/20) of this package for [rOpenSci](https://ropensci.org/).
