### 2015-10-17 8278743
Create corpus from character vector
Fixes #44
