+++ R/TextReuseCorpus.R
@@ -4,12 +4,19 @@
 #' virtual S3 class \code{\link[tm]{Corpus}} from the \code{tm} package. The
 #' object is a \code{TextReuseCorpus}, which is basically a list containing
 #' objects of class \code{\link{TextReuseTextDocument}}. Arguments are passed
-#' along to that constructor function. You can pass either a character vector of
-#' paths to text files using the \code{paths =} parameter, or a directory
-#' containing text files using the \code{dir =} parameter.
+#' along to that constructor function. To create the corpus, you can pass either
+#' a character vector of paths to text files using the \code{paths =} parameter,
+#' a directory containing text files (with any extension) using the \code{dir =}
+#' parameter, or a character vector of documents using the \code{text = }
+#' parameter, where each element in the characer vector is a document. If the
+#' character vector passed to \code{text = } has names, then those names will be
+#' used as the document IDs. Otherwise, IDs will be assigned to the documents.
+#' Only one of the \code{paths}, \code{dir}, or \code{text} parameters should be
+#' specified.
 #'
 #' @param paths A character vector of paths to files to be opened.
 #' @param dir A The path to a directory of text files.
+#' @param text A character vector (possibly named) of documents.
 #' @param meta A list with named elements for the metadata associated with this
 #'   corpus.
 #' @param progress Display a progress bar while loading files.
@@ -31,18 +38,11 @@
 #' names(corpus)
 #' corpus[["ca1851-match"]]
 #' @export
-TextReuseCorpus <- function(paths, dir = NULL, meta = list(),
+TextReuseCorpus <- function(paths, dir = NULL, text = NULL, meta = list(),
                             progress = interactive(),
                             tokenizer = tokenize_ngrams, ...,
                             hash_func = hash_string,
                             keep_tokens = FALSE, keep_text = TRUE) {
-
-  if (missing(paths) & !is.null(dir)) {
-    assert_that(is.dir(dir))
-    paths <- Sys.glob(str_c(dir, "/*"))
-  }
-
-  vapply(paths, is.readable, logical(1), USE.NAMES = FALSE)
 
   if (!is.null(tokenizer)) {
     assert_that(is.function(tokenizer),
@@ -56,24 +56,68 @@ TextReuseCorpus <- function(paths, dir = NULL, meta = list(),
     loading_msg <- "Loading "
   }
 
-  if (progress) {
-    len <- length(paths)
-    message(loading_msg, prettyNum(len, big.mark = ","), " documents.")
-    pb <- txtProgressBar(min = 0, max = len, style = 3)
+  # If we get a character vector of documents, use that; otherwise load
+  # the files from disk.
+  if (!missing(text)) {
+
+    assert_that(missing(paths),
+                is.null(dir),
+                is.character(text))
+
+    if (progress) {
+      len <- length(text)
+      message(loading_msg, prettyNum(len, big.mark = ","), " documents.")
+      pb <- txtProgressBar(min = 0, max = len, style = 3)
+    }
+
+    if (is.null(names(text)))
+      names(text) <- str_c("doc-", 1:length(text))
+
+    docs <- lapply(seq_along(text), function(i) {
+      d <- TextReuseTextDocument(text = text[i], tokenizer = tokenizer, ...,
+                                 hash_func = hash_func,
+                                 keep_tokens = keep_tokens,
+                                 keep_text = keep_text,
+                                 meta = list(id = names(text)[i],
+                                             tokenizer = tokenizer_name,
+                                             hash_func = hash_func_name))
+      if (progress) setTxtProgressBar(pb, i)
+      d
+    })
+
+    if (progress) close(pb)
+
[…]
