### 2015-10-17 f3f0945
Fix type
Fixes #42
