### 2015-10-17 b6d7207
Skip short or empty documents
Fixes #45
