### 2015-10-17 47300bf
Tokenize and rehash write names to meta
Fixes #38
