+++ R/tokenize.R
@@ -15,8 +15,15 @@
 #'   returned or discarded?
 #' @param keep_text Should the text be saved in the document that is returned or
 #'   discarded?
+#'
 #' @return The modified \code{\link{TextReuseTextDocument}} or
 #'   \code{\link{TextReuseCorpus}}.
+#'
+#' @examples
+#' dir <- system.file("extdata/legal", package = "textreuse")
+#' corpus <- TextReuseCorpus(dir = dir, tokenizer = NULL)
+#' corpus <- tokenize(corpus, tokenize_ngrams)
+#' head(tokens(corpus[[1]]))
 #' @export
 tokenize <- function(x, tokenizer, ..., hash_func = hash_string,
                      keep_tokens = FALSE, keep_text = TRUE) {
@@ -35,6 +42,8 @@ tokenize.TextReuseTextDocument <- function(x, tokenizer, ...,
   x$hashes <- hash_func(x$tokens)
   if (!keep_tokens) x$tokens <- NULL
   if (!keep_text) x$text <- NULL
+  x$meta$tokenizer <- as.character(substitute(tokenizer))
+  x$meta$hash_func <- as.character(substitute(hash_func))
   x
 }
 
@@ -44,5 +53,10 @@ tokenize.TextReuseCorpus <- function(x, tokenizer, ..., hash_func = hash_string,
   x$documents <- lapply(x$documents, tokenize, tokenizer, ...,
                         hash_func = hash_func, keep_tokens = keep_tokens,
                         keep_text = keep_text)
+
+  x$meta$tokenizer <- as.character(substitute(tokenizer))
+  x$meta$hash_func <- as.character(substitute(hash_func))
+
   x
+
 }
