### 2015-10-17 453b410
Replace column checks with class identifier
Fixes #46
