+++ R/pairwise_candidates.R
@@ -30,6 +30,6 @@ pairwise_candidates <- function(m, directional = FALSE) {
   df <- data.frame(a = a, b = b, score = score, stringsAsFactors = FALSE)
   if (!directional) df <- sort_df_by_rows(df)
   df <- sort_df_by_columns(df)
-  class(df) <- c("tbl_df", "tbl", "data.frame")
+  class(df) <- c("tbl_df", "tbl", "data.frame", "textreuse_candidates")
   df
 }
