### 2015-10-17 06194dc
Option to check short documents
Fixes #48
