+++ R/TextReuseTextDocument.R
@@ -21,6 +21,7 @@
 #'   returned or discarded?
 #' @param keep_text Should the text be saved in the document that is returned or
 #'   discarded?
+#' @param skip_short Should short documents be skipped? (See details.)
 #'
 #' @details This constructor function follows a three-step process. It reads in
 #'   the text, either from a file or from memory. It then tokenizes that text.
@@ -29,12 +30,12 @@
 #'   \code{keep_tokens} and \code{keep_text}, you can avoid saving those
 #'   objects, which can result in significant memory savings for large corpora.
 #'
-#'   This constructor function will return \code{NULL} for very short or empty
-#'   documents. A very short document is one where there are two few words to
-#'   create at least two n-grams. For example, if five-grams are desired, then a
-#'   document must be at least six words long. If no value of \code{n} is
-#'   provided, then the function assumes a value of \code{n = 3}. A warning will
-#'   be printed with the document ID of a skipped document.
+#'   If \code{skip_short = TRUE}, this function will return \code{NULL} for very
+#'   short or empty documents. A very short document is one where there are two
+#'   few words to create at least two n-grams. For example, if five-grams are
+#'   desired, then a document must be at least six words long. If no value of
+#'   \code{n} is provided, then the function assumes a value of \code{n = 3}. A
+#'   warning will be printed with the document ID of a skipped document.
 #'
 #' @return An object of class \code{TextReuseTextDocument}. This object inherits
 #'   from the virtual S3 class \code{\link[NLP]{TextDocument}} in the NLP
@@ -57,7 +58,9 @@
 TextReuseTextDocument <- function(text, file = NULL, meta = list(),
                                   tokenizer = tokenize_ngrams, ...,
                                   hash_func = hash_string,
-                                  keep_tokens = FALSE, keep_text = TRUE) {
+                                  keep_tokens = FALSE,
+                                  keep_text = TRUE,
+                                  skip_short = TRUE) {
 
   if (!missing(text)) assert_that(has_id(meta))
 
@@ -71,15 +74,17 @@ TextReuseTextDocument <- function(text, file = NULL, meta = list(),
   text <- as_string(text)
 
   # Check length of document
-  n_call <- match.call(expand.dots = TRUE)[["n"]]
-  if (is.null(n_call))
-    n_call <- 3
-  if (wordcount(text) < n_call + 1) {
-    warning("Skipping document with ID '", meta$id,
-            "' because it has too few words ",
-            "to create at least two n-grams with n = ", n_call, ".",
-            call. = FALSE, noBreaks. = TRUE)
-    return(NULL)
+  if (skip_short) {
+    n_call <- match.call(expand.dots = TRUE)[["n"]]
+    if (is.null(n_call))
+      n_call <- 3
+    if (wordcount(text) < n_call + 1) {
+      warning("Skipping document with ID '", meta$id,
+              "' because it has too few words ",
+              "to create at least two n-grams with n = ", n_call, ".",
+              call. = FALSE, noBreaks. = TRUE)
+      return(NULL)
+    }
   }
 
   if (!is.null(tokenizer)) {
