+++ man/TextReuseCorpus.Rd
@@ -7,7 +7,8 @@
 \usage{
 TextReuseCorpus(paths, dir = NULL, text = NULL, meta = list(),
   progress = interactive(), tokenizer = tokenize_ngrams, ...,
-  hash_func = hash_string, keep_tokens = FALSE, keep_text = TRUE)
+  hash_func = hash_string, keep_tokens = FALSE, keep_text = TRUE,
+  skip_short = TRUE)
 
 is.TextReuseCorpus(x)
 }
@@ -56,12 +57,12 @@ Only one of the \code{paths}, \code{dir}, or \code{text} parameters should be
 specified.
 }
 \details{
-This constructor function will skip very short or empty documents. A
-  very short document is one where there are two few words to create at least
-  two n-grams. For example, if five-grams are desired, then a document must
-  be at least six words long. If no value of \code{n} is provided, then the
-  function assumes a value of \code{n = 3}. A warning will be printed with
-  the document ID of each skipped document.
+If \code{skip_short = TRUE}, this function will skip very short or
+  empty documents. A very short document is one where there are two few words
+  to create at least two n-grams. For example, if five-grams are desired,
+  then a document must be at least six words long. If no value of \code{n} is
+  provided, then the function assumes a value of \code{n = 3}. A warning will
+  be printed with the document ID of each skipped document.
 }
 \examples{
 dir <- system.file("extdata/legal", package = "textreuse")
