### 2015-10-18 81aea23
Reimplement LSH with dplr
Squashed commit of the following:

commit a0c1ed420de4b34fafde7ef3125fc55ab7d7aaa3
Author: Lincoln Mullen <lincoln@lincolnmullen.com>
Date:   Sat Oct 17 21:00:38 2015 -0400

    Implement LSH with dplyr

    Fixes #47

commit 1b8dd4ec4c5faa435b1c6309ac8f436dbb72acca
Author: Lincoln Mullen <lincoln@lincolnmullen.com>
Date:   Sat Oct 17 17:47:35 2015 -0400

    Initial reimplementation of LSH with dplyr

commit d768a433271c76081bacc341e6b8db1526377193
Author: Lincoln Mullen <lincoln@lincolnmullen.com>
Date:   Sat Oct 17 17:42:27 2015 -0400

    Option to check short doc
[…]
