+++ R/utils.R
@@ -11,12 +11,12 @@ pretty_print_metadata <- function(doc) {
 }
 
 # Check whether the number of minhashes is evenly divisible by number of bands
-check_banding <- function(l, b) {
-  l %% b == 0
+check_banding <- function(h, b) {
+  h %% b == 0
 }
 
 assertthat::on_failure(check_banding) <- function(call, env) {
-  "Bands times rows must equal the number of hashes."
+  "The number of hashes must be evenly divisible by the number of bands."
 }
 
 # Sequences for subsetting by bands in minhash
@@ -25,15 +25,6 @@ band_seq <- function(l, b) {
   r <- l / b
   starts <- seq.int(from = 1, to = l, by = r)
   lapply(starts, function(n) seq.int(n, n + r - 1, 1))
-}
-
-# Append to value if key already exists
-insert_into_hash <- function(k, v, hash) {
-  if (hash::has.key(k, hash))
-    hash[[k]] <- unique(c(hash[[k]], v))
-  else
-    hash[[k]] <- v
-  invisible()
 }
 
 # Test that meta exists and that it has an ID value
@@ -45,6 +36,20 @@ assertthat::on_failure(has_id) <- function(call, env) {
   paste("When creating a document from a string instead of a file, the `id`",
         "field in the metadata list must be specified.")
 }
+
+# People might row_bind() two of these data frames, so we can't rely just on
+# the class.
+is_lsh_buckets <- function(x) {
+  class_check <- inherits(x, "lsh_buckets")
+  col_check <- identical(names(x), c("doc", "buckets")) &
+    inherits(x, "data.frame")
+  class_check | col_check
+}
+
+assertthat::on_failure(is_lsh_buckets) <- function(call, env) {
+  "Object is not a data frame of LSH buckets."
+}
+
 
 is_candidates_df <- function(x) {
   inherits(x, "textreuse_candidates")
