### 2015-10-18 #20 @lmullen
I've implemented almost all of your suggestions, @noamross. They are now on the master branch of the repository, or you can get the most recent version tagged `v0.0.1.9004`. Here is a list of the most significant changes:

- As mentioned above, LSH is now implemented using dplyr. For my [ATS corpus](http://lincolnmullen.com/blog/corpus-of-american-tract-society-publications/) I got exactly the same results with the new implementation as with the old implementation, and I tested an earlier version on a different corpus, so I'm pretty confident that it is correct.
- You can now created a corpus using a character vector of documents, and if the character vector is named, the names will be used as the document IDs.
- Creating a document or a corpus now checks whether the document has at least `n + 1` words, where `n` is length of the n-grams. (And by extension, it checks for empty documents.) When creating a corpus, you get a warning with the ID of empty or short documents but the corpus just skips over time.

The one suggestion that you made that I did not implement is somehow storing or at least validating the minhash function when creating adding new documents to the LSH buckets/cache. I agree in principle that this should be done, but I couldn't think of a good way to do it. I don't think it is a big issue for the first version of the package, since the package can only deal with in-memory data at this point. That is to say, the biggest corpus that I'm likely to use it on has 71K documents and that isn't even close to using up the memory on my laptop. In those cases, it is unlikely that someone would add to the cache instead of simply doing it all in one batch. For the future I'm thinking about how this could be extended for out of memory data, which the new dplyr implementation will permit. I think I'll have to figure out how to implement your suggestion then.

Thanks again for your detailed review. It has improved the package a great deal.
