### 2015-10-21 3ab9daf
Compute Smith-Waterman matrix with Rcpp
