+++ NAMESPACE
@@ -4,7 +4,6 @@ S3method("[",TextReuseCorpus)
 S3method("[[",TextReuseCorpus)
 S3method("content<-",TextReuseTextDocument)
 S3method("hashes<-",TextReuseTextDocument)
-S3method("meta<-",TextReuseCorpus)
 S3method("meta<-",TextReuseTextDocument)
 S3method("names<-",TextReuseCorpus)
 S3method("tokens<-",TextReuseTextDocument)
@@ -20,7 +19,6 @@ S3method(jaccard_similarity,default)
 S3method(length,TextReuseCorpus)
 S3method(lsh,TextReuseCorpus)
 S3method(lsh,TextReuseTextDocument)
-S3method(meta,TextReuseCorpus)
 S3method(meta,TextReuseTextDocument)
 S3method(names,TextReuseCorpus)
 S3method(print,TextReuseCorpus)
@@ -39,9 +37,11 @@ S3method(wordcount,default)
 export("content<-")
 export("hashes<-")
 export("meta<-")
+export("meta<-.TextReuseCorpus")
 export("tokens<-")
 export(TextReuseCorpus)
 export(TextReuseTextDocument)
+export(align_local)
 export(content)
 export(filenames)
 export(has_content)
@@ -61,6 +61,7 @@ export(lsh_probability)
 export(lsh_subset)
 export(lsh_threshold)
 export(meta)
+export(meta.TextReuseCorpus)
 export(minhash_generator)
 export(pairwise_candidates)
 export(pairwise_compare)
