### 2015-10-21 48b4767
Better warnings about parameters
