### 2015-10-21 773f221
Infrastructure for traceback function
