+++ src/sw_traceback.cpp
@@ -1,0 +1,7 @@
+#include <Rcpp.h>
+using namespace Rcpp;
+
+// [[Rcpp::export]]
+IntegerMatrix sw_traceback(IntegerMatrix m) {
+  return m;
+}
