### 2015-10-21 0c9a2a4
Skip locale specific tests on Windows
