+++ tests/testthat/test-wordcount.R
@@ -12,6 +12,7 @@ test_that("counts words correctly for different classes", {
 
 test_that("counts words for a corpus", {
   skip_on_appveyor()
+  skip_on_os("windows")
   wc <- wordcount(corpus)
   expect_true(!is.null(names(wc)))
   wc <- unname(wc)
