### 2015-10-21 236d523
Imports from stats and utils
