+++ R/align_local.R
@@ -1,20 +1,25 @@
 #' @export
+# http://etherealbits.com/2013/04/string-alignment-dynamic-programming-dna/
 align_local <- function(a, b, tokenizer = tokenize_words, ...,
-                        match = 2L, mismatch = -1L, gap = -1L) {
+                        match = 2L, mismatch = -1L, gap = -1L,
+                        delete_mark = "#") {
 
   assert_that(is.character(a),
               is.character(b),
               is_integer_like(match),
               is_integer_like(mismatch),
-              is_integer_like(gap))
+              is_integer_like(gap),
+              is.string(delete_mark))
 
-  if (match <= 0 || mismatch > 0 || gap > 0) {
+  if (match <= 0 || mismatch > 0 || gap > 0 || !(nchar(delete_mark) == 1)) {
     stop("The scoring parameters should have the following characteristics:\n",
          "    - `match` should be a positive integer\n",
          "    - `mismatch` should be a negative integer or zero\n",
-         "    - `gap` should be a negative integer or zero\n")
+         "    - `gap` should be a negative integer or zero\n",
+         "    - `delete_mark` should be a single character\n")
   }
 
+  # Keep everything as integers because IntegerMatrix saves memory
   match    <- as.integer(match)
   mismatch <- as.integer(mismatch)
   gap      <- as.integer(gap)
@@ -31,13 +36,103 @@ align_local <- function(a, b, tokenizer = tokenize_words, ...,
 
   # Create the integer matrix
   m <- matrix(0L, length(b) + 1L, length(a) + 1L)
-  rownames(m) <- c(NA, b)
-  colnames(m) <- c(NA, a)
+  # rownames(m) <- c(NA, b)
+  # colnames(m) <- c(NA, a)
 
+  # Calculate the matrix of possible paths
   m <- sw_matrix(m, a, b, match, mismatch, gap)
 
-  m <- sw_traceback(m)
+  # Find the starting place in the matrix
+  max_match <- which(m == max(m), arr.ind = TRUE, useNames = FALSE)
 
-  m
+  if (nrow(max_match) > 1) {
+    message("Multiple best alignments found; selecting only one of them.")
+  }
+
+  # Create output vectors which are as long as conceivably necessary
+  a_out <- vector(mode = "character", length = max(max_match))
+  b_out <- vector(mode = "character", length = max(max_match))
+  a_out[] <- NA_character_
+  b_out[] <- NA_character_
+
+  # Initialize counters for the matrix and the output vector
+  row_i <- max_match[1, 1]
+  col_i <- max_match[1, 2]
+  out_i <- 1L
+
+  # Place our first known values in the output vectors
+  b_out[out_i] <- b[row_i - 1]
+  a_out[out_i] <- a[col_i - 1]
+  out_i = out_i + 1L # Advance the out vector position
+
+  # Begin moving up, left, or diagonally within the matrix till we hit a zero
+  while (m[row_i, col_i] != 0) {
+
+    # Values of the current cell, the cells up, left, diagonal, and the max
+    up       <- m[row_i - 1, col_i]
+    left     <- m[row_i, col_i - 1]
+    diagn    <- m[row_i - 1, col_i - 1]
+    max_cell <- max(up, left, diagn)
+
+    # Move in the direction of the maximum cell. If there are ties, choose up
+    # first, then left, then diagonal. Privilege up and left because they
+    # preserve edits.
+    #
+    # In each case add the current words to the out vectors. For moves up and
+    # and left there will be an insertion/deletion, so add a symbol like ####
+    # that is the same number of characters as the word in the other vector.
+    #
[…]
