### 2015-10-21 38d24bf
Preserve capitalization in output
