### 2015-10-21 85295e3
Make local_align an S3 method
