+++ NAMESPACE
@@ -8,6 +8,8 @@ S3method("meta<-",TextReuseCorpus)
 S3method("meta<-",TextReuseTextDocument)
 S3method("names<-",TextReuseCorpus)
 S3method("tokens<-",TextReuseTextDocument)
+S3method(align_local,TextReuseTextDocument)
+S3method(align_local,default)
 S3method(as.character,TextReuseTextDocument)
 S3method(content,TextReuseTextDocument)
 S3method(hashes,TextReuseCorpus)
