### 2015-10-21 3dc16b3
Fix bug where we moved too far back in the SW matrix
