### 2015-10-21 1abb9d7
Progress bar and interrupt for local align
