+++ DESCRIPTION
@@ -20,10 +20,10 @@ Imports: assertthat (>= 0.1),
     dplyr (>= 0.4.3),
     NLP (>= 0.1.8),
     Rcpp (>= 0.12.0),
+    RcppProgress  (>= 0.1),
     stringr (>= 1.0.0),
     tidyr (>= 0.3.1)
 Suggests: testthat (>= 0.11.0),
     knitr (>= 1.11),
     rmarkdown (>= 0.8)
-LinkingTo: Rcpp,
-    BH
+LinkingTo: BH, Rcpp, RcppProgress
