### 2015-10-21 23cdc9b
Print method for TextReuse alignment objects
