+++ NAMESPACE
@@ -27,6 +27,7 @@ S3method(meta,TextReuseTextDocument)
 S3method(names,TextReuseCorpus)
 S3method(print,TextReuseCorpus)
 S3method(print,TextReuseTextDocument)
+S3method(print,textreuse_alignment)
 S3method(ratio_of_matches,TextReuseTextDocument)
 S3method(ratio_of_matches,default)
 S3method(rehash,TextReuseCorpus)
