### 2015-10-21 b2dbba4
Decide on progress bar before creating matrix
