### 2015-10-21 b8cbc17
Add an alignment score to output
