### 2015-10-21 047c46e
Add Smith-Waterman to package-level documentation
