+++ DESCRIPTION
@@ -8,7 +8,9 @@ Authors@R: c(person("Lincoln", "Mullen", role = c("aut", "cre"),
 Description: Tools for measuring similarity among documents and detecting
     passages which have been reused. Implements shingled n-gram, skip n-gram,
     and other tokenizers; similarity/dissimilarity functions; pairwise
-    comparisons; and minhash and locality sensitive hashing algorithms.
+    comparisons; minhash and locality sensitive hashing algorithms; and a
+    version of the Smith-Waterman local alignment algorithm suitable for
+    natural language.
 License: MIT + file LICENSE
 LazyData: TRUE
 URL: https://github.com/lmullen/textreuse
