### 2015-10-21 3fe0a05
Better progress messages
