### 2015-10-22 dadf720
Declare import from RcppProgress
