+++ NAMESPACE
@@ -77,6 +77,7 @@ export(tokenize_skip_ngrams)
 export(tokenize_words)
 export(tokens)
 export(wordcount)
+import(RcppProgress)
 import(assertthat)
 import(stringr)
 importFrom(NLP,"content<-")
