### 2015-10-22 651830b
Documentation for align_local
