### 2015-10-22 f703eee
Keep list in the class
