### 2015-10-22 f7f6a16
Tests for alignment
