### 2015-10-22 bea9d0c
Prefix for all vignettes
