### 2015-10-22 b703429
Vignette on Smith-Waterman
