+++ R/textreuse-package.r
@@ -18,6 +18,8 @@
 #'
 #' \code{vignette("textreuse-minhash", package = "textreuse")}
 #'
+#' \code{vignette("textreuse-alignment", package = "textreuse")}
+#'
 #' Another good place to beign with the package is the documentation for loading
 #' documents (\code{\link{TextReuseTextDocument}} and
 #' \code{\link{TextReuseCorpus}}), for \link{tokenizers},
