+++ DESCRIPTION
@@ -1,8 +1,8 @@
 Package: textreuse
 Type: Package
 Title: Detect Text Reuse and Document Similarity
-Version: 0.0.1.9004
-Date: 2015-09-22
+Version: 0.0.1.9005
+Date: 2015-10-21
 Authors@R: c(person("Lincoln", "Mullen", role = c("aut", "cre"),
     email = "lincoln@lincolnmullen.com"))
 Description: Tools for measuring similarity among documents and detecting
