### 2015-10-22 #20 @lmullen
I've added one last feature (I promise this time!). It is a local alignment function that implements a version of the Smith-Waterman function from protein alignment for use with natural language. It is one function `local_align()`, there is a vignette `textreuse-alignment` to go with it, and the new tag is `v0.0.1.9005`.
