+++ man/lsh.Rd
@@ -57,7 +57,7 @@ dir <- system.file("extdata/legal", package = "textreuse")
 minhash <- minhash_generator(200, seed = 235)
 corpus <- TextReuseCorpus(dir = dir,
                           tokenizer = tokenize_ngrams, n = 5,
-                          hash_func = minhash)
+                          minhash_func = minhash)
 buckets <- lsh(corpus, bands = 50)
 buckets
 }
