+++ man/lsh_subset.Rd
@@ -22,7 +22,7 @@ dir <- system.file("extdata/legal", package = "textreuse")
 minhash <- minhash_generator(200, seed = 234)
 corpus <- TextReuseCorpus(dir = dir,
                           tokenizer = tokenize_ngrams, n = 5,
-                          hash_func = minhash)
+                          minhash_func = minhash)
 buckets <- lsh(corpus, bands = 50)
 candidates <- lsh_candidates(buckets)
 lsh_subset(candidates)

[…]
