+++ R/lsh_subset.R
@@ -9,7 +9,7 @@
 #' minhash <- minhash_generator(200, seed = 234)
 #' corpus <- TextReuseCorpus(dir = dir,
 #'                           tokenizer = tokenize_ngrams, n = 5,
-#'                           hash_func = minhash)
+#'                           minhash_func = minhash)
 #' buckets <- lsh(corpus, bands = 50)
 #' candidates <- lsh_candidates(buckets)
 #' lsh_subset(candidates)
