### 2015-10-23 0829e7f
Parallelize embarrasingly parallel problems
