+++ vignettes/textreuse-introduction.Rmd
@@ -143,3 +143,17 @@ See the documentation for `?similarity-functions` for details on what is measure
 
 You can write your own similarity functions, which should accept two sets or bags, `a` and `b`, should work on both character and numeric vectors, since they are used with either tokens or hashes of tokens, and should return a single numeric score for the comparison. You will need to implement a method for the `TextReuseTextDocument` class.
 
+## Parallelization
+
+This package will use multiple cores for a few functions is an option is set. This only benefits the corpus loading and tokenizing functions, which are often the slowest parts of an analysis. This is implemented with the [parallel package](https://cran.r-project.org/web/views/HighPerformanceComputing.html), and does not work on Windows machines. (Regardless of the options set, this package will never attempt to parallelize computations on Windows.)
+
+To use the parallel option, you must specify the number of CPU cores that you wish to use:
+
+```{R eval = FALSE}
+options("mc.cores" = 4L)
+```
+
+If that option is set, this package will use multiple cores when possible.
+
+You can figure out how many cores your computer has with `parallel::detectCores()`. See `help(package = "parallel")` for more details.
+
