### 2015-10-25 fc00e75
New function lsh_query()
