+++ R/lsh.R
@@ -47,7 +47,8 @@
 #'   for Dummies}" (November 14, 2013).
 #'
 #'@seealso \code{\link{minhash_generator}}, \code{\link{lsh_candidates}},
-#'  \code{\link{lsh_probability}}, \code{\link{lsh_threshold}}
+#'  \code{\link{lsh_query}}, \code{\link{lsh_probability}},
+#'  \code{\link{lsh_threshold}}
 #'
 #' @examples
 #' dir <- system.file("extdata/legal", package = "textreuse")
