+++ NAMESPACE
@@ -67,6 +67,7 @@ export(lsh)
 export(lsh_candidates)
 export(lsh_compare)
 export(lsh_probability)
+export(lsh_query)
 export(lsh_subset)
 export(lsh_threshold)
 export(meta)
