### 2015-10-25 24503bc
New function lsh_query()
