### 2015-10-25 54b1b44
As matrix method
