### 2015-10-25 72c53b1
As matrix method
