+++ tests/testthat/test-candidate_pairs.R
@@ -8,5 +8,12 @@ test_that("returns a data frame with correct properties", {
   expect_equal(nrow(pairs), 6)
   expect_is(pairs, "data.frame")
   expect_is(pairs, "tbl_df")
+  expect_is(pairs, "textreuse_candidates")
   expect_named(pairs, c("a", "b", "score"))
+})
+
+test_that("can be converted to matrix", {
+  m <- as.matrix(pairs)
+  expect_is(m, "matrix")
+  expect_equal(names(corpus), colnames(m))
 })
