+++ NAMESPACE
@@ -12,6 +12,7 @@ S3method("tokens<-",TextReuseTextDocument)
 S3method(align_local,TextReuseTextDocument)
 S3method(align_local,default)
 S3method(as.character,TextReuseTextDocument)
+S3method(as.matrix,textreuse_candidates)
 S3method(content,TextReuseTextDocument)
 S3method(hashes,TextReuseCorpus)
 S3method(hashes,TextReuseTextDocument)
