### 2015-10-28 d0aa6a5
Only define string for hasing once
