### 2015-10-28 #20 @noamross
*Note: Github's "compare" functionality is super helpful for doing this review. I just did `https://github.com/lmullen/textreuse/compare/f9d3555...HEAD` to find the updated changes.*

-  All tests pass on R 3.2.2, OSX El Capitan,
-  Code of conduct added; 👍
- Loading text via charactervector works great;  👍
- Nice package-level documentation and documentation updates; 👍
-  New LSH algorithm looks good;  👍
-  When loading from a directory of files including some to be skipped via `skip_short = TRUE`, the warning messages do not give an ID for the document. Thus, you can't tell which might be the problematic file. I see when I load a named character vector there is an ID in the warning message. I believe this is because the ID is assigned after the check for the short document in `TextReuseTextDocument()`.

- I believe that for your Rcpp function `sw_matrix`, you want to define `score`, `deletion`, `insertion`, and `value` *before* the for loop, and only assign to them....ah, it's easier to just do a PR: https://github.com/lmullen/textreuse/pull/59

- I'm still grokking the new Smith-Waterman algorithm, but I wonder if we could test by comparing results with a bioinformatics package that implements it.  If we created a test    file where the only words are space-separated "A", "C", "T", and "G" we should expect identical results, no?   So far I haven't been able to do this with `Biostrings::pairwiseAlignment()`, but I'm not sure how to set the score assignments in that properly.

    Perhaps @vsbuffalo, @blahah404 or one of our other bioinfo friends could chime in with the right usage for expected results? I'm trying to do `Biostrings::pairwiseAlignment()`, with type="local" and set scores for matching (2), and mismatch/gap (-1).
