### 2015-10-28 93a8766
Fix error messages with missing IDs
Fixes #57
Fixes #55
