### 2015-10-28 c5fc3c3
Fix bug with incorrect capitalization of last word of local alignment
