### 2015-10-28 #20 @lmullen
Thanks for doing the second thorough review, @noamross. Much appreciated!

Here are the responses to your queries/notes:

- The missing ID was definitely a bug. Fixed in [this commit](https://github.com/lmullen/textreuse/commit/93a876604cff3e7851ca377c10c45056296a4693), with tests added.

- Thanks for the PR for `sw_matrix()`. I've merged it in. (Related: @Ironholds sent a [PR](https://github.com/lmullen/textreuse/pull/61) to improve `hash_string()`.

- The Biostrings and the textreuse implementations of Smith-Waterman can't exactly be compared directly, since Biostrings permits substitutions, where I've opted to mark substitutions as a deletion then an insertion. The reasons is that there is a limited vocabulary that Biostrings is working with, but there is an unlimited vocabulary for natural language. But I've prepared [this document](http://rpubs.com/lmullen/textreuse-cf-to-biostrings) which demonstrates that they get the same results.

In addition, I've cleared out any remaining issues in the repository.

@sckott, @karthik: Anything else to be done before this package can be part of rOpenSci?
