+++ man/lsh_query.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/lsh_query.R
 \name{lsh_query}
 \alias{lsh_query}
@@ -25,6 +25,7 @@ corpus <- TextReuseCorpus(dir = dir,
                           minhash_func = minhash)
 buckets <- lsh(corpus, bands = 50)
 lsh_query(buckets, "ny1850-match")
+
 }
 \seealso{
 \code{\link{lsh}}, \code{\link{lsh_candidates}}
