+++ man/similarity-functions.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/similarity.R
 \name{similarity-functions}
 \alias{jaccard_bag_similarity}
@@ -21,7 +21,7 @@ ratio_of_matches(a, b)
 directional comparisons.}
 
 \item{b}{The second set (or bag) to be compared. The destination bag for
-  directional comparisons.}
+directional comparisons.}
 }
 \description{
 A set of functions which take two sets or bag of words and measure their
@@ -89,6 +89,7 @@ ratio_of_matches(ny, ca_match)
 # These two should have lower similarity scores
 jaccard_similarity(ny, ca_nomatch)
 ratio_of_matches(ny, ca_nomatch)
+
 }
 \references{
 Jure Leskovec, Anand Rajaraman, and Jeff Ullman,
