+++ man/lsh_candidates.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/lsh_candidates.R
 \name{lsh_candidates}
 \alias{lsh_candidates}
@@ -24,5 +24,6 @@ corpus <- TextReuseCorpus(dir = dir,
                           minhash_func = minhash)
 buckets <- lsh(corpus, bands = 50)
 lsh_candidates(buckets)
+
 }
 
