+++ man/align_local.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/align_local.R
 \name{align_local}
 \alias{align_local}
@@ -27,7 +27,7 @@ negative integer or zero.}
 insertions/deletions in the documents.}
 
 \item{progress}{Display a progress bar and messages while computing the
-  alignment.}
+alignment.}
 }
 \value{
 A list with the class \code{textreuse_alignment}. This list contains
@@ -73,6 +73,7 @@ dir <- system.file("extdata/legal", package = "textreuse")
 corpus <- TextReuseCorpus(dir = dir, progress = FALSE)
 alignment <- align_local(corpus[["ca1851-match"]], corpus[["ny1850-match"]])
 str(alignment)
+
 }
 \references{
 For a useful description of the algorithm, see
