+++ man/TextReuseCorpus.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/TextReuseCorpus.R
 \name{TextReuseCorpus}
 \alias{TextReuseCorpus}
@@ -80,6 +80,7 @@ corpus <- TextReuseCorpus(dir = dir, meta = list("description" = "Field Codes"))
 corpus[[1]]
 names(corpus)
 corpus[["ca1851-match"]]
+
 }
 \seealso{
 \link[=TextReuseTextDocument-accessors]{Access for TextReuse
