+++ man/lsh_subset.Rd
@@ -1,4 +1,4 @@
-% Generated by roxygen2 (4.1.1): do not edit by hand
+% Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/lsh_subset.R
 \name{lsh_subset}
 \alias{lsh_subset}
