+++ NAMESPACE
@@ -93,7 +93,6 @@ importFrom(NLP,"meta<-")
 importFrom(NLP,content)
 importFrom(NLP,meta)
 importFrom(Rcpp,sourceCpp)
-importFrom(stats,runif)
 importFrom(utils,getTxtProgressBar)
 importFrom(utils,setTxtProgressBar)
 importFrom(utils,txtProgressBar)
