### 2015-10-29 63b81cd
Redo imported/exported functions
