+++ man/NLP-imports.Rd
@@ -1,12 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/TextReuseTextDocument.R
-\name{meta}
-\alias{content}
-\alias{content<-}
-\alias{meta}
-\alias{meta<-}
-\title{Accessors for TextReuseTextDocument objects}
-\description{
-The \code{meta} and \code{content} generics are defined by the NLP package. See the \code{\link[NLP]{generics}} documentation in that package.
-}
-
