+++ man/reexports.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/TextReuseTextDocument.R
+\docType{import}
+\name{reexports}
+\alias{content}
+\alias{content<-}
+\alias{meta}
+\alias{meta<-}
+\alias{reexports}
+\title{Objects exported from other packages}
+\description{
+These objects are imported from other packages. Follow the links
+below to see their documentation.
+
+\describe{
+  \item{NLP}{\code{\link[NLP]{meta}}, \code{\link[NLP]{meta<-}}, \code{\link[NLP]{content}}, \code{\link[NLP]{content<-}}}
+}}
+\keyword{internal}
+
