### 2015-10-29 ee3aac4
Redo package documentation
