+++ man/textreuse-package.Rd
@@ -1,17 +1,17 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/textreuse-package.r
 \docType{package}
-\name{textreuse}
+\name{textreuse-package}
 \alias{textreuse}
 \alias{textreuse-package}
-\title{textreuse: Detect Text Reuse and Document Similarity}
+\title{Detect Text Reuse and Document Similarity}
 \description{
-This R package provides a set of functions for measuring similarity among
-documents and detecting passages which have been reused. It implements
-shingled n-gram, skip n-gram, and other tokenizers; similarity/dissimilarity
-functions; pairwise comparisons; minhash and locality sensitive hashing
-algorithms, and a version of the Smith-Waterman local alignment algorithm
-suitable for natural language.
+Tools for measuring similarity among documents and detecting
+passages which have been reused. Implements shingled n-gram, skip n-gram,
+and other tokenizers; similarity/dissimilarity functions; pairwise
+comparisons; minhash and locality sensitive hashing algorithms; and a
+version of the Smith-Waterman local alignment algorithm suitable for
+natural language.
 }
 \details{
 The best place to begin with this package in the introductory vignette.
