### 2015-10-29 #20 @sckott
LGTM, `approved`

@lmullen i think you can move pkgs into ropensci already
