### 2015-10-30 f776cdd
Soften check for candidates data frame
