### 2015-10-30 7f4b2e2
rOpenSci footer
