+++ README.Rmd
@@ -22,8 +22,8 @@ An R package for detecting text reuse and document similarity in a corpus.
 **Author:** [Lincoln Mullen](http://lincolnmullen.com)<br>
 **License:** [MIT](http://opensource.org/licenses/MIT)<br>
 
-[![Build Status](https://travis-ci.org/lmullen/textreuse.svg?branch=master)](https://travis-ci.org/lmullen/textreuse) 
-[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/lmullen/textreuse?branch=master)](https://ci.appveyor.com/project/lmullen/textreuse)
+[![Build Status](https://travis-ci.org/ropensci/textreuse.svg?branch=master)](https://travis-ci.org/lmullen/textreuse) 
+[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/textreuse?branch=master)](https://ci.appveyor.com/project/lmullen/textreuse)
 
 ### Description 
 
@@ -35,7 +35,7 @@ This package is not yet on CRAN. To install it from GitHub, use [devtools](https
 
 ```{r eval=FALSE}
 # install.packages("devtools")
-devtools::install_github("lmullen/textreuse", build_vignettes = TRUE)
+devtools::install_github("ropensci/textreuse", build_vignettes = TRUE)
 ```
 
 ### Examples
