### 2015-10-30 #20 @lmullen
Okay, I've transferred the repository and added the rOpenSci footer, etc.
